functions/file_retrieval/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
Flask==2.*